refactor: RacingGame Class형 변경, 분리
